Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/html/semantics/document-metadata/the-base-element/base_href_invalid-expected.txt


FAIL base element with unparseable href should have .href getter return attr value assert_equals: expected "//test:test" but got ""
